feat: Reorganizing code structure
