--- api/api.go
@@ -1,12 +1,0 @@
-package api
-
-type TicketRequest struct {
-	Name   string
-	Phone  string
-	PaxNum int
-}
-
-type TicketResponse struct {
-	QueueId          string
-	QueuePositionNum int
-}
